React to changes to root's child list
Previously, headings would sometimes fail to be linked if the page loads directly into a storylet. This was because the "router-example" div is constructed while the loading screen displays, and is only attached as a child of the "root" div after it is fully constructed. The mutations which introduce the storylet elements therefore occur without the MutationObserver being able to see them.

The solution is to have the MutationObserver additionally watch for changes to the "root" div's own child list.