Make links open in a new tab
Fallen London faithfully restores page state on reload _most_ of the time, but not _all_ of the time. Certain ephemeral views – such as the outcome of a storylet – are lost if the user navigates away from the page. This can be especially frustrating if the user was attempting to read the wiki page on the storylet before reading its final resolution. Therefore, to improve UX, we should not cause users to navigate away from the page.